requirements.txt
spotipy
pyjwt
requests